Improve SDK Resolver (Description for more info)
SDK would try to load on every imported function. We cache the pointer to the library and reuse that. This may cause exceptions if the library is unloaded for any reason. We have added a config option in the `CorsairSDK.AdvancedOptions` to turn caching off.

Exceptions will no longer be thrown in the SDK Resolver.

Improved Signed Certificate checks for the Corsair Native SDK